requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
streamlit==1.38.0
orjson==3.10.7
google-generativeai==0.8.5
